requirements.txt
django
django-rest-framework
dj-rest-auth
django-allauth
requests